requirement.txt
python=3.6.8
tensorflow-gpu=1.11.0(tensorflow=1.11.0)
numpy=1.14.0
gensim=3.8.1